Restic: Add restic and replace vwbackup
This commit adds the restic backup image to automatically perform
incremental backups of important folders replacing the old vwbackup
custom image.
This substitution is done to decouple the backup from vaultwarden since
in the future it's likely that we need to backup other data too. This
way we have a general-purpose tool that backs-up all kind of important
data.
At the moment the only data backed up is vaultwarden's data folder.

As part of the migration to restic we had some changes in the bitwarden
folder removind the vaultwarden and vwbackup subfolders. The same this
happened with the compose file, where only the vaultwarden service has
remained.