Caddy: Move log and block requests outside LAN
This commit moves the log block from the global config inside the site
block in the `Caddyfile`; this is done in order to properly log each
request received by the domain. The logs will be read by crowdsec to
determine malevolent access like DDOS.

Another chage is the addition of a snippet used to block all accesses to
bitwarden admin panel from outside the local LAN. Directly blocking each
route from outside is better than redirecting to the / route and can be
improved in the future adding more private routes.