Networks: Rename shared networks
This commit renames two networks that are shared between containers:
`caddy` and `notification`.
The networks are used to connect a central container that act as a
server or proxy to other user containers so it's natural that the
network is created by the central container through compose (like all
the other networks).
The new networks are named: `caddy_default` and `notify_default`.