Caddy: Make internal services accesible with proxy
This commit makes the services portainer and ddns-updater accessible
through the caddy reverse proxy.
Previously we needed to open ports for them in order to display the
WebUI, but now with caddy it's just like bitwarden.
Those services can be accessed only inside the local network, but in the
future we could have some service accessible from the outside (like a
personal website) and some in parti pubblic and in part private.